refactor(ff-filter): split filter_inner/mod.rs into focused submodules
