+++ crates/ff-filter/src/filter_inner/mod.rs
@@ -14,22 +14,12 @@
 
 mod build;
 mod convert;
+mod normalize;
+mod push_pull;
 
-use build::{
-    add_and_link_step, add_atempo_chain, add_fit_to_aspect_pad, add_join_with_dissolve_step,
-    add_overlay_image_step, add_parametric_eq_chain, add_raw_filter_step, add_setpts_after_trim,
-    audio_buffersrc_args, create_hw_filter, hw_accel_to_device_type, video_buffersrc_args,
-};
-use convert::{
-    audio_pts_ticks, av_frame_to_audio_frame, av_frame_to_video_frame, copy_audio_planes_to_av,
-    copy_video_planes_to_av, pixel_format_to_av, sample_format_to_av, sample_format_to_av_name,
-    video_pts_ticks,
-};
-
-use std::os::raw::c_int;
 use std::ptr::NonNull;
 
-use ff_format::{AudioFrame, VideoFrame};
+use ff_format::AudioFrame;
 
 use crate::error::FilterError;
 use crate::graph::{FilterStep, HwAccel};
@@ -176,1526 +166,13 @@ impl FilterGraphInner {
             peak_pass2_done: false,
         }
     }
-
-    // ── Video ─────────────────────────────────────────────────────────────────
-
-    /// Lazily initialise the video filter graph from the first pushed frame.
-    fn ensure_video_graph(&mut self, frame: &VideoFrame) -> Result<(), FilterError> {
-        if self.graph.is_some() {
-            return Ok(());
-        }
-
-        let pix_fmt = pixel_format_to_av(frame.format());
-        let args = video_buffersrc_args(frame.width(), frame.height(), pix_fmt);
-        let num_inputs = self.video_input_count();
-
-        // SAFETY: all raw pointers are checked for null after allocation; the
-        // graph pointer is stored in `self.graph` and kept alive for the
-        // lifetime of this struct.
-        unsafe {
-            let graph_ptr = ff_sys::avfilter_graph_alloc();
-            if graph_ptr.is_null() {
-                return Err(FilterError::BuildFailed);
-            }
-            // SAFETY: checked non-null above.
-            let graph_nn = NonNull::new_unchecked(graph_ptr);
-
-            match Self::build_video_graph(
-                graph_nn,
-                &args,
-                num_inputs,
-                &self.steps,
-                self.hw.as_ref(),
-            ) {
-                Ok((src_ctxs, vsink_ctx, hw_device_ctx)) => {
-                    self.graph = Some(graph_nn);
-                    self.src_ctxs = src_ctxs;
-                    self.vsink_ctx = Some(vsink_ctx);
-                    self.hw_device_ctx = hw_device_ctx;
-                    log::info!(
-                        "filter graph configured inputs={} filters={}",
-                        num_inputs,
-                        self.steps.len()
-                    );
-                    Ok(())
-                }
-                Err(e) => {
-                    let mut raw = graph_nn.as_ptr();
-                    ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(raw));
-                    Err(e)
-                }
-            }
-        }
-    }
-
-    /// Build the `AVFilterGraph` for video, returning `(src_ctxs, vsink_ctx)`.
-    ///
-    /// `num_inputs` buffersrc contexts are created (`in0`..`inN-1`).  For
-    /// multi-input filters like `overlay`, the extra sources are linked to the
-    /// appropriate input pads after the main chain link is established.
-    ///
-    /// # Safety
-    ///
-    /// `graph_nn` must be a valid, freshly-allocated `AVFilterGraph`.
-    unsafe fn build_video_graph(
-        graph_nn: NonNull<ff_sys::AVFilterGraph>,
-        buffersrc_args: &str,
-        num_inputs: usize,
-        steps: &[FilterStep],
-        hw: Option<&HwAccel>,
-    ) -> VideoGraphResult {
-        let graph = graph_nn.as_ptr();
-
-        // 0. When hardware acceleration is requested, create a device context
-        //    and disable automatic pixel-format conversion so FFmpeg does not
-        //    insert implicit hwupload/scale filters that would conflict with the
-        //    explicit ones we add below.
-        let hw_device_ctx: Option<*mut ff_sys::AVBufferRef> = if let Some(hw) = hw {
-            let device_type = hw_accel_to_device_type(*hw);
-            let mut raw_hw_ctx: *mut ff_sys::AVBufferRef = std::ptr::null_mut();
-            let ret = ff_sys::av_hwdevice_ctx_create(
-                &raw mut raw_hw_ctx,
-                device_type,
-                std::ptr::null(),     // device: null = system default
-                std::ptr::null_mut(), // opts: null = defaults
-                0,
-            );
-            if ret < 0 {
-                log::warn!("av_hwdevice_ctx_create failed hw={hw:?} code={ret}");
-                return Err(FilterError::BuildFailed);
-            }
-            // AVFILTER_AUTO_CONVERT_NONE = 0: hardware filters must receive
-            // frames in exactly the format they expect.
-            ff_sys::avfilter_graph_set_auto_convert(graph, 0u32);
-            log::debug!("hw device context created hw={hw:?}");
-            Some(raw_hw_ctx)
-        } else {
-            None
-        };
-
-        // Helper closure: free hw_device_ctx and return an error.  Used at
-        // every early-return failure point that occurs *after* the device
-        // context has been allocated so it is not leaked.
-        macro_rules! bail {
-            ($err:expr) => {{
-                if let Some(mut hw_ctx) = hw_device_ctx {
-                    ff_sys::av_buffer_unref(std::ptr::addr_of_mut!(hw_ctx));
-                }
-                return Err($err);
-            }};
-        }
-
-        // SAFETY: `avfilter_get_by_name` returns a borrowed pointer valid for
-        // the process lifetime; we never free it.
-        let buffersrc = ff_sys::avfilter_get_by_name(c"buffer".as_ptr());
-        if buffersrc.is_null() {
-            bail!(FilterError::BuildFailed);
-        }
-
-        let Ok(src_args) = std::ffi::CString::new(buffersrc_args) else {
-            bail!(FilterError::BuildFailed)
-        };
-        let mut src_ctxs: Vec<Option<NonNull<ff_sys::AVFilterContext>>> =
-            Vec::with_capacity(num_inputs);
-
-        // 1. Create in0 (always present).
-        let mut raw_ctx0: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-        let ret = ff_sys::avfilter_graph_create_filter(
-            &raw mut raw_ctx0,
-            buffersrc,
-            c"in0".as_ptr(),
-            src_args.as_ptr(),
-            std::ptr::null_mut(),
-            graph,
-        );
-        if ret < 0 {
-            bail!(FilterError::BuildFailed);
-        }
-        log::debug!("filter added name=buffersrc slot=0");
-        // SAFETY: ret >= 0 guarantees non-null.
-        src_ctxs.push(Some(NonNull::new_unchecked(raw_ctx0)));
-
-        // Create in1..inN-1 (for overlay etc.)
-        for slot in 1..num_inputs {
-            let Ok(ctx_name) = std::ffi::CString::new(format!("in{slot}")) else {
-                bail!(FilterError::BuildFailed)
-            };
-            let mut raw_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-            let ret = ff_sys::avfilter_graph_create_filter(
-                &raw mut raw_ctx,
-                buffersrc,
-                ctx_name.as_ptr(),
-                src_args.as_ptr(),
-                std::ptr::null_mut(),
-                graph,
-            );
-            if ret < 0 {
-                bail!(FilterError::BuildFailed);
-            }
-            log::debug!("filter added name=buffersrc slot={slot}");
-            // SAFETY: ret >= 0 guarantees non-null.
-            src_ctxs.push(Some(NonNull::new_unchecked(raw_ctx)));
-        }
-
-        // 2. Create buffersink ("buffersink").
-        let buffersink = ff_sys::avfilter_get_by_name(c"buffersink".as_ptr());
-        if buffersink.is_null() {
-            bail!(FilterError::BuildFailed);
-        }
-
-        let mut sink_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-        let ret = ff_sys::avfilter_graph_create_filter(
-            &raw mut sink_ctx,
-            buffersink,
-            c"out".as_ptr(),
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-            graph,
-        );
-        if ret < 0 {
-            bail!(FilterError::BuildFailed);
-        }
-
-        // 3. Insert hwupload/hwupload_cuda BEFORE the filter steps so that
-        //    subsequent filters receive hardware (CUDA/VAAPI/VTB) frames.
-        let mut prev_ctx: *mut ff_sys::AVFilterContext = raw_ctx0;
-        if let (Some(hw_ctx), Some(hw_backend)) = (hw_device_ctx, hw) {
-            let upload_name = match hw_backend {
-                HwAccel::Cuda => c"hwupload_cuda",
-                HwAccel::VideoToolbox | HwAccel::Vaapi => c"hwupload",
-            };
-            prev_ctx = match create_hw_filter(graph, prev_ctx, upload_name, c"hwupload0", hw_ctx) {
-                Ok(ctx) => ctx,
-                Err(e) => bail!(e),
-            };
-        }
-
-        // 4-5. Add each `FilterStep`, link the main chain (in0 → step[0] → …),
-        // and wire extra input pads for multi-input filters.
-        for (i, step) in steps.iter().enumerate() {
-            // Audio-only steps; skip them in the video graph.
-            if matches!(
-                step,
-                FilterStep::AReverse
-                    | FilterStep::AFadeIn { .. }
-                    | FilterStep::AFadeOut { .. }
-                    | FilterStep::ParametricEq { .. }
-                    | FilterStep::ANoiseGate { .. }
-                    | FilterStep::ACompressor { .. }
-                    | FilterStep::StereoToMono
-                    | FilterStep::ChannelMap { .. }
-                    | FilterStep::AudioDelay { .. }
-                    | FilterStep::ConcatAudio { .. }
-            ) {
-                continue;
-            }
-
-            // LoudnessNormalize is audio-only and handled via two-pass in
-            // push_audio / pull_audio rather than through the filter graph.
-            if matches!(step, FilterStep::LoudnessNormalize { .. }) {
-                continue;
-            }
-
-            // NormalizePeak is audio-only and handled via two-pass buffering in
-            // push_audio / pull_audio.
-            if matches!(step, FilterStep::NormalizePeak { .. }) {
-                continue;
-            }
-
-            // OverlayImage is a compound step (movie → lut → overlay).  It
-            // creates its own internal source node via the `movie` filter and
-            // does not consume a buffersrc slot, so it must bypass the standard
-            // `add_and_link_step` path which assumes a single filter per step.
-            if let FilterStep::OverlayImage {
-                path,
-                x,
-                y,
-                opacity,
-            } = step
-            {
-                prev_ctx = match add_overlay_image_step(graph, prev_ctx, path, x, y, *opacity, i) {
-                    Ok(ctx) => ctx,
-                    Err(e) => bail!(e),
-                };
-                continue;
-            }
-
-            // JoinWithDissolve is a compound step that expands to:
-            //   prev → trim_a → setpts_a → xfade[0]
-            //   in1  → trim_b → setpts_b → xfade[1]
-            // It bypasses the standard add_and_link_step path entirely.
-            if let FilterStep::JoinWithDissolve {
-                clip_a_end,
-                clip_b_start,
-                dissolve_dur,
-            } = step
-            {
-                let Some(b_src) = src_ctxs.get(1).and_then(|o| *o) else {
-                    bail!(FilterError::BuildFailed)
-                };
-                prev_ctx = match add_join_with_dissolve_step(
-                    graph,
-                    prev_ctx,
-                    b_src.as_ptr(),
-                    *clip_a_end,
-                    *clip_b_start,
-                    *dissolve_dur,
-                    i,
-                ) {
-                    Ok(ctx) => ctx,
-                    Err(e) => bail!(e),
-                };
-                continue;
-            }
-
-            prev_ctx = match add_and_link_step(graph, prev_ctx, step, i, "step") {
-                Ok(ctx) => ctx,
-                Err(e) => bail!(e),
-            };
-
-            // After trim, insert setpts=PTS-STARTPTS so the output timestamps
-            // are reset to start at zero.
-            if matches!(step, FilterStep::Trim { .. }) {
-                prev_ctx = match add_setpts_after_trim(graph, prev_ctx, i) {
-                    Ok(ctx) => ctx,
-                    Err(e) => bail!(e),
-                };
-            }
-
-            // FitToAspect is a compound step: the scale filter (added by
-            // add_and_link_step above) preserves the source aspect ratio; the
-            // pad filter added here centres the scaled frame on the target canvas.
-            if let FilterStep::FitToAspect {
-                width,
-                height,
-                color,
-            } = step
-            {
-                prev_ctx = match add_fit_to_aspect_pad(graph, prev_ctx, *width, *height, color, i) {
-                    Ok(ctx) => ctx,
-                    Err(e) => bail!(e),
-                };
-            }
-
-            // Overlay and xfade both consume a second input on pad 1.
-            if matches!(step, FilterStep::Overlay { .. } | FilterStep::XFade { .. })
-                && let Some(Some(extra_src)) = src_ctxs.get(1)
-            {
-                let ret = ff_sys::avfilter_link(extra_src.as_ptr(), 0, prev_ctx, 1);
-                if ret < 0 {
-                    bail!(FilterError::BuildFailed);
-                }
-                log::debug!("filter linked extra_input=in1 to two-input filter pad=1");
-            }
-
-            // ConcatVideo consumes n input pads; link src_ctxs[1..n-1] to pads 1..n-1.
-            if let FilterStep::ConcatVideo { n } = step {
-                for slot in 1..*n as usize {
-                    if let Some(Some(extra_src)) = src_ctxs.get(slot) {
-                        let ret =
-                            ff_sys::avfilter_link(extra_src.as_ptr(), 0, prev_ctx, slot as u32);
-                        if ret < 0 {
-                            bail!(FilterError::BuildFailed);
-                        }
-                        log::debug!("filter linked extra_input=in{slot} to concat pad={slot}");
-                    }
-                }
-            }
-        }
-
-        // 6. Insert hwdownload AFTER all filter steps so output frames are
-        //    downloaded back to system memory before reaching the buffersink.
-        if let Some(hw_ctx) = hw_device_ctx {
-            prev_ctx =
-                match create_hw_filter(graph, prev_ctx, c"hwdownload", c"hwdownload0", hw_ctx) {
-                    Ok(ctx) => ctx,
-                    Err(e) => bail!(e),
-                };
-        }
-
-        // Link last filter to sink.
-        let ret = ff_sys::avfilter_link(prev_ctx, 0, sink_ctx, 0);
-        if ret < 0 {
-            bail!(FilterError::BuildFailed);
-        }
-
-        // 7. Configure the graph.
-        let ret = ff_sys::avfilter_graph_config(graph, std::ptr::null_mut());
-        if ret < 0 {
-            log::warn!("avfilter_graph_config failed code={ret}");
-            // If there is a crop step the most likely cause is the rectangle
-            // extending beyond the source frame dimensions.
-            if let Some(FilterStep::Crop {
-                x,
-                y,
-                width,
-                height,
-            }) = steps.iter().find(|s| matches!(s, FilterStep::Crop { .. }))
-            {
-                bail!(FilterError::InvalidConfig {
-                    reason: format!(
-                        "crop rect {x},{y}+{width}x{height} exceeds source frame dimensions"
-                    ),
-                });
-            }
-            bail!(ffmpeg_err(ret));
-        }
-
-        // SAFETY: `avfilter_graph_create_filter` with ret >= 0 guarantees
-        // non-null pointers.
-        let sink_nn = NonNull::new_unchecked(sink_ctx);
-        Ok((src_ctxs, sink_nn, hw_device_ctx))
-    }
-
-    /// Push a video frame into the filter graph.
-    pub(crate) fn push_video(
-        &mut self,
-        slot: usize,
-        frame: &VideoFrame,
-    ) -> Result<(), FilterError> {
-        self.ensure_video_graph(frame)?;
-
-        let src_ctx = self
-            .src_ctxs
-            .get(slot)
-            .and_then(|opt| *opt)
-            .ok_or_else(|| FilterError::InvalidInput {
-                slot,
-                reason: format!("slot {slot} out of range (have {})", self.src_ctxs.len()),
-            })?;
-
-        // SAFETY: we allocate the `AVFrame`, fill it with `VideoFrame` data,
-        // push it to the buffersrc, then immediately free it.  The buffersrc
-        // keeps its own reference (`AV_BUFFERSRC_FLAG_KEEP_REF`).
-        unsafe {
-            let raw_frame = ff_sys::av_frame_alloc();
-            if raw_frame.is_null() {
-                return Err(FilterError::ProcessFailed);
-            }
-
-            (*raw_frame).width = frame.width() as c_int;
-            (*raw_frame).height = frame.height() as c_int;
-            (*raw_frame).format = pixel_format_to_av(frame.format());
-            let pts = video_pts_ticks(frame.timestamp());
-            if pts == ff_sys::AV_NOPTS_VALUE {
-                log::warn!("pts invalid, passing AV_NOPTS_VALUE to filter graph slot={slot}");
-            }
-            (*raw_frame).pts = pts;
-
-            let ret = ff_sys::av_frame_get_buffer(raw_frame, 0);
-            if ret < 0 {
-                let mut ptr = raw_frame;
-                ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-                return Err(FilterError::ProcessFailed);
-            }
-
-            copy_video_planes_to_av(frame, raw_frame);
-
-            let ret = ff_sys::av_buffersrc_add_frame_flags(
-                src_ctx.as_ptr(),
-                raw_frame,
-                ff_sys::BUFFERSRC_FLAG_KEEP_REF,
-            );
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-
-            if ret < 0 {
-                return Err(FilterError::ProcessFailed);
-            }
-        }
-        Ok(())
-    }
-
-    /// Pull the next filtered video frame, or `None` if not yet available.
-    pub(crate) fn pull_video(&mut self) -> Result<Option<VideoFrame>, FilterError> {
-        let Some(sink_ctx) = self.vsink_ctx else {
-            return Ok(None);
-        };
-
-        // SAFETY: we allocate a temporary `AVFrame`, hand it to
-        // `av_buffersink_get_frame`, convert the result, then free it.
-        unsafe {
-            let raw_frame = ff_sys::av_frame_alloc();
-            if raw_frame.is_null() {
-                return Err(FilterError::ProcessFailed);
-            }
-
-            let ret = ff_sys::av_buffersink_get_frame(sink_ctx.as_ptr(), raw_frame);
-
-            // EAGAIN (-11) and EOF: return `None`.
-            if ret < 0 {
-                let mut ptr = raw_frame;
-                ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-                return Ok(None);
-            }
-
-            let result = av_frame_to_video_frame(raw_frame);
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-
-            match result {
-                Ok(frame) => Ok(Some(frame)),
-                Err(()) => Err(FilterError::ProcessFailed),
-            }
-        }
-    }
-
-    // ── Audio ─────────────────────────────────────────────────────────────────
-
-    /// Returns the number of video input slots required by the configured steps.
-    ///
-    /// Returns 2 when [`FilterStep::Overlay`], [`FilterStep::XFade`], or
-    /// [`FilterStep::JoinWithDissolve`] is present (each needs a main stream on
-    /// slot 0 and a secondary stream on slot 1), 1 otherwise.
-    fn video_input_count(&self) -> usize {
-        for step in &self.steps {
-            if matches!(
-                step,
-                FilterStep::Overlay { .. }
-                    | FilterStep::XFade { .. }
-                    | FilterStep::JoinWithDissolve { .. }
-            ) {
-                return 2;
-            }
-            if let FilterStep::ConcatVideo { n } = step {
-                return *n as usize;
-            }
-        }
-        1
-    }
-
-    /// Returns the number of audio input slots required by the configured steps.
-    fn audio_input_count(&self) -> usize {
-        for step in &self.steps {
-            if let FilterStep::Amix(n) = step {
-                return *n;
-            }
-            if let FilterStep::ConcatAudio { n } = step {
-                return *n as usize;
-            }
-        }
-        1
-    }
-
-    /// Lazily initialise the audio filter graph from the first pushed frame.
-    fn ensure_audio_graph(&mut self, frame: &AudioFrame) -> Result<(), FilterError> {
-        if self.asink_ctx.is_some() {
-            return Ok(());
-        }
-
-        let num_inputs = self.audio_input_count();
-        let sample_fmt = sample_format_to_av_name(frame.format());
-        let sample_rate = frame.sample_rate();
-        let channels = frame.channels();
-
-        let args = audio_buffersrc_args(sample_rate, sample_fmt, channels);
-
-        // SAFETY: same contract as `ensure_video_graph` — pointers checked for
-        // null, stored in `self`, freed in `Drop`.
-        unsafe {
-            let graph_ptr = ff_sys::avfilter_graph_alloc();
-            if graph_ptr.is_null() {
-                return Err(FilterError::BuildFailed);
-            }
-            // SAFETY: checked non-null above.
-            let graph_nn = NonNull::new_unchecked(graph_ptr);
-
-            match Self::build_audio_graph(
-                graph_nn,
-                &args,
-                num_inputs,
-                &self.steps,
-                self.hw.as_ref(),
-            ) {
-                Ok((src_ctxs, asink_ctx)) => {
-                    if self.graph.is_none() {
-                        self.graph = Some(graph_nn);
-                    } else {
-                        let mut raw = graph_nn.as_ptr();
-                        ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(raw));
-                    }
-                    let video_slots = self.src_ctxs.len();
-                    self.src_ctxs.resize(video_slots + num_inputs, None);
-                    for (i, ctx) in src_ctxs.into_iter().enumerate() {
-                        self.src_ctxs[video_slots + i] = ctx;
-                    }
-                    self.asink_ctx = Some(asink_ctx);
-                    log::info!(
-                        "filter graph configured inputs={} filters={}",
-                        num_inputs,
-                        self.steps.len()
-                    );
-                    Ok(())
-                }
-                Err(e) => {
-                    let mut raw = graph_nn.as_ptr();
-                    ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(raw));
-                    Err(e)
-                }
-            }
-        }
-    }
-
-    /// Build the `AVFilterGraph` for audio, returning `(src_ctxs, asink_ctx)`.
-    ///
-    /// # Safety
-    ///
-    /// `graph_nn` must be a valid, freshly-allocated `AVFilterGraph`.
-    unsafe fn build_audio_graph(
-        graph_nn: NonNull<ff_sys::AVFilterGraph>,
-        buffersrc_args: &str,
-        num_inputs: usize,
-        steps: &[FilterStep],
-        _hw: Option<&HwAccel>,
-    ) -> BuildResult {
-        let graph = graph_nn.as_ptr();
-        let mut src_ctxs = Vec::with_capacity(num_inputs);
-
-        // 1. Create abuffer sources, one per input slot.
-        let abuffer = ff_sys::avfilter_get_by_name(c"abuffer".as_ptr());
-        if abuffer.is_null() {
-            return Err(FilterError::BuildFailed);
-        }
-        let src_args =
-            std::ffi::CString::new(buffersrc_args).map_err(|_| FilterError::BuildFailed)?;
-
-        // First input slot.
-        let first_src_ctx = {
-            let mut raw_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-            let ret = ff_sys::avfilter_graph_create_filter(
-                &raw mut raw_ctx,
-                abuffer,
-                c"in0".as_ptr(),
-                src_args.as_ptr(),
-                std::ptr::null_mut(),
-                graph,
-            );
-            if ret < 0 {
-                return Err(FilterError::BuildFailed);
-            }
-            log::debug!("filter added name=abuffersrc slot=0");
-            // SAFETY: ret >= 0 means raw_ctx is non-null.
-            let nn = NonNull::new_unchecked(raw_ctx);
-            src_ctxs.push(Some(nn));
-            nn
-        };
-
-        // Additional input slots for amix.
-        for slot in 1..num_inputs {
-            let ctx_name = std::ffi::CString::new(format!("in{slot}"))
-                .map_err(|_| FilterError::BuildFailed)?;
-            let mut raw_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-            let ret = ff_sys::avfilter_graph_create_filter(
-                &raw mut raw_ctx,
-                abuffer,
-                ctx_name.as_ptr(),
-                src_args.as_ptr(),
-                std::ptr::null_mut(),
-                graph,
-            );
-            if ret < 0 {
-                return Err(FilterError::BuildFailed);
-            }
-            log::debug!("filter added name=abuffersrc slot={slot}");
-            // SAFETY: ret >= 0 means raw_ctx is non-null.
-            src_ctxs.push(Some(NonNull::new_unchecked(raw_ctx)));
-        }
-
-        // 2. Create abuffersink.
-        let abuffersink = ff_sys::avfilter_get_by_name(c"abuffersink".as_ptr());
-        if abuffersink.is_null() {
-            return Err(FilterError::BuildFailed);
-        }
-
-        let mut sink_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-        let ret = ff_sys::avfilter_graph_create_filter(
-            &raw mut sink_ctx,
-            abuffersink,
-            c"aout".as_ptr(),
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-            graph,
-        );
-        if ret < 0 {
-            return Err(FilterError::BuildFailed);
-        }
-
-        // 3-5. Add each `FilterStep` (audio-relevant steps) and link.
-        let mut prev_ctx = first_src_ctx.as_ptr();
-        for (i, step) in steps.iter().enumerate() {
-            // Video-only steps; skip them in the audio graph.
-            if matches!(
-                step,
-                FilterStep::Reverse
-                    | FilterStep::ConcatVideo { .. }
-                    | FilterStep::JoinWithDissolve { .. }
-            ) {
-                continue;
-            }
-
-            // LoudnessNormalize is handled via two-pass buffering in
-            // push_audio / pull_audio; the regular audio graph is never built
-            // when this step is present, so this guard is defensive only.
-            if matches!(step, FilterStep::LoudnessNormalize { .. }) {
-                continue;
-            }
-
-            // NormalizePeak is handled via two-pass buffering in
-            // push_audio / pull_audio; same reasoning as LoudnessNormalize.
-            if matches!(step, FilterStep::NormalizePeak { .. }) {
-                continue;
-            }
-
-            // Speed uses `setpts` for video but `atempo` for audio.  Bypass the
-            // standard `add_and_link_step` path and insert the atempo chain here.
-            if let FilterStep::Speed { factor } = step {
-                prev_ctx = add_atempo_chain(graph, prev_ctx, *factor, i)?;
-                continue;
-            }
-
-            // ParametricEq generates one filter node per band; bypass the
-            // single-node `add_and_link_step` path.
-            if let FilterStep::ParametricEq { bands } = step {
-                prev_ctx = add_parametric_eq_chain(graph, prev_ctx, bands, i)?;
-                continue;
-            }
-
-            // AudioDelay dispatches to adelay (positive/zero) or atrim (negative).
-            if let FilterStep::AudioDelay { ms } = step {
-                let (filter_name, args) = if *ms >= 0.0 {
-                    ("adelay".to_string(), format!("delays={ms}:all=1"))
-                } else {
-                    ("atrim".to_string(), format!("start={}", -ms / 1000.0))
-                };
-                // SAFETY: graph and prev_ctx are valid pointers in the same graph.
-                prev_ctx = add_raw_filter_step(graph, prev_ctx, &filter_name, &args, i, "adelay")?;
-                continue;
-            }
-
-            prev_ctx = add_and_link_step(graph, prev_ctx, step, i, "astep")?;
-
-            // ConcatAudio consumes n input pads; link src_ctxs[1..n-1] to pads 1..n-1.
-            if let FilterStep::ConcatAudio { n } = step {
-                for slot in 1..*n as usize {
-                    if let Some(Some(extra_src)) = src_ctxs.get(slot) {
-                        let ret =
-                            ff_sys::avfilter_link(extra_src.as_ptr(), 0, prev_ctx, slot as u32);
-                        if ret < 0 {
-                            return Err(FilterError::BuildFailed);
-                        }
-                        log::debug!("filter linked extra_input=in{slot} to concat pad={slot}");
-                    }
-                }
-            }
-        }
-
-        // Link last filter to sink.
-        let ret = ff_sys::avfilter_link(prev_ctx, 0, sink_ctx, 0);
-        if ret < 0 {
-            return Err(FilterError::BuildFailed);
-        }
-
-        // 6. Configure the graph.
-        let ret = ff_sys::avfilter_graph_config(graph, std::ptr::null_mut());
-        if ret < 0 {
-            log::warn!("avfilter_graph_config failed code={ret}");
-            return Err(ffmpeg_err(ret));
-        }
-
-        // SAFETY: sink_ctx is non-null (ret >= 0 above).
-        let sink_nn = NonNull::new_unchecked(sink_ctx);
-        Ok((src_ctxs, sink_nn))
-    }
-
-    /// Push an audio frame into the filter graph.
-    pub(crate) fn push_audio(
-        &mut self,
-        slot: usize,
-        frame: &AudioFrame,
-    ) -> Result<(), FilterError> {
-        // Two-pass loudness normalization: buffer frames instead of feeding the
-        // graph.  The measurement + correction passes run on the first pull_audio.
-        if self
-            .steps
-            .iter()
-            .any(|s| matches!(s, FilterStep::LoudnessNormalize { .. }))
-        {
-            self.loudness_buf.push(frame.clone());
-            return Ok(());
-        }
-
-        // Two-pass peak normalization: buffer frames instead of feeding the graph.
-        if self
-            .steps
-            .iter()
-            .any(|s| matches!(s, FilterStep::NormalizePeak { .. }))
-        {
-            self.peak_buf.push(frame.clone());
-            return Ok(());
-        }
-
-        self.ensure_audio_graph(frame)?;
-
-        let audio_inputs = self.audio_input_count();
-        let video_slots = self.src_ctxs.len().saturating_sub(audio_inputs);
-        let audio_slot = video_slots + slot;
-
-        let src_ctx = self
-            .src_ctxs
-            .get(audio_slot)
-            .and_then(|opt| *opt)
-            .ok_or_else(|| FilterError::InvalidInput {
-                slot,
-                reason: format!("audio slot {slot} out of range (have {audio_inputs})"),
-            })?;
-
-        // SAFETY: allocate `AVFrame`, copy `AudioFrame` data, push, free.
-        unsafe {
-            let raw_frame = ff_sys::av_frame_alloc();
-            if raw_frame.is_null() {
-                return Err(FilterError::ProcessFailed);
-            }
-
-            (*raw_frame).nb_samples = frame.samples() as c_int;
-            (*raw_frame).sample_rate = frame.sample_rate() as c_int;
-            (*raw_frame).format = sample_format_to_av(frame.format());
-            let pts = audio_pts_ticks(frame.timestamp(), frame.sample_rate());
-            if pts == ff_sys::AV_NOPTS_VALUE {
-                log::warn!("pts invalid, passing AV_NOPTS_VALUE to filter graph slot={slot}");
-            }
-            (*raw_frame).pts = pts;
-            (*raw_frame).ch_layout.nb_channels = frame.channels() as c_int;
-
-            let ret = ff_sys::av_frame_get_buffer(raw_frame, 0);
-            if ret < 0 {
-                let mut ptr = raw_frame;
-                ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-                return Err(FilterError::ProcessFailed);
-            }
-
-            copy_audio_planes_to_av(frame, raw_frame);
-
-            let ret = ff_sys::av_buffersrc_add_frame_flags(
-                src_ctx.as_ptr(),
-                raw_frame,
-                ff_sys::BUFFERSRC_FLAG_KEEP_REF,
-            );
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-
-            if ret < 0 {
-                return Err(FilterError::ProcessFailed);
-            }
-        }
-        Ok(())
-    }
-
-    /// Pull the next filtered audio frame, or `None` if not yet available.
-    pub(crate) fn pull_audio(&mut self) -> Result<Option<AudioFrame>, FilterError> {
-        // Two-pass loudness normalization: run both passes on the first call,
-        // then drain the corrected output frame-by-frame.
-        if self
-            .steps
-            .iter()
-            .any(|s| matches!(s, FilterStep::LoudnessNormalize { .. }))
-        {
-            if !self.loudness_pass2_done {
-                self.run_loudness_normalization()?;
-            }
-            if self.loudness_output_idx < self.loudness_output.len() {
-                let frame = self.loudness_output[self.loudness_output_idx].clone();
-                self.loudness_output_idx += 1;
-                return Ok(Some(frame));
-            }
-            return Ok(None);
-        }
-
-        // Two-pass peak normalization: run both passes on the first call,
-        // then drain the corrected output frame-by-frame.
-        if self
-            .steps
-            .iter()
-            .any(|s| matches!(s, FilterStep::NormalizePeak { .. }))
-        {
-            if !self.peak_pass2_done {
-                self.run_peak_normalization()?;
-            }
-            if self.peak_output_idx < self.peak_output.len() {
-                let frame = self.peak_output[self.peak_output_idx].clone();
-                self.peak_output_idx += 1;
-                return Ok(Some(frame));
-            }
-            return Ok(None);
-        }
-
-        let Some(sink_ctx) = self.asink_ctx else {
-            return Ok(None);
-        };
-
-        // SAFETY: allocate, fill via `av_buffersink_get_frame`, convert, free.
-        unsafe {
-            let raw_frame = ff_sys::av_frame_alloc();
-            if raw_frame.is_null() {
-                return Err(FilterError::ProcessFailed);
-            }
-
-            let ret = ff_sys::av_buffersink_get_frame(sink_ctx.as_ptr(), raw_frame);
-
-            // EAGAIN (-11) and EOF: return `None`.
-            if ret < 0 {
-                let mut ptr = raw_frame;
-                ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-                return Ok(None);
-            }
-
-            let result = av_frame_to_audio_frame(raw_frame);
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-
-            match result {
-                Ok(frame) => Ok(Some(frame)),
-                Err(()) => Err(FilterError::ProcessFailed),
-            }
-        }
-    }
-
-    // ── Two-pass loudness normalization ──────────────────────────────────────
-
-    /// Run EBU R128 two-pass loudness normalization over `self.loudness_buf`:
-    ///
-    /// 1. Measure integrated loudness with an `ebur128=peak=true:metadata=1` graph.
-    /// 2. Compute `gain_db = target_lufs − measured_lufs`.
-    /// 3. Apply gain with a `volume={gain_db}dB` graph.
-    /// 4. Store corrected frames in `self.loudness_output`.
-    fn run_loudness_normalization(&mut self) -> Result<(), FilterError> {
-        let target_lufs = self
-            .steps
-            .iter()
-            .find_map(|s| {
-                if let FilterStep::LoudnessNormalize { target_lufs, .. } = s {
-                    Some(*target_lufs)
-                } else {
-                    None
-                }
-            })
-            .ok_or(FilterError::BuildFailed)?;
-
-        // Mark done early to prevent re-entry on error.
-        self.loudness_pass2_done = true;
-
-        if self.loudness_buf.is_empty() {
-            return Ok(());
-        }
-
-        // === Pass 1: measure integrated loudness ===
-        let measured_lufs = unsafe {
-            let graph = ff_sys::avfilter_graph_alloc();
-            if graph.is_null() {
-                return Err(FilterError::BuildFailed);
-            }
-            let result = run_ebur128_graph(graph, &self.loudness_buf);
-            let mut g = graph;
-            ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(g));
-            result?
-        };
-
-        let gain_db = target_lufs - measured_lufs;
-        log::info!(
-            "loudness normalization measured_lufs={:.1} target_lufs={:.1} gain_db={:.2}",
-            measured_lufs,
-            target_lufs,
-            gain_db,
-        );
-
-        // === Pass 2: apply volume correction ===
-        self.loudness_output = unsafe {
-            let graph = ff_sys::avfilter_graph_alloc();
-            if graph.is_null() {
-                return Err(FilterError::BuildFailed);
-            }
-            let result = run_volume_graph(graph, &self.loudness_buf, gain_db);
-            let mut g = graph;
-            ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(g));
-            result?
-        };
-
-        Ok(())
-    }
-
-    // ── Two-pass peak normalization ───────────────────────────────────────────
-
-    /// Run peak-level two-pass normalization over `self.peak_buf`:
-    ///
-    /// 1. Measure the true peak with an `astats=metadata=1` graph.
-    /// 2. Compute `gain_db = target_db − measured_peak_db`.
-    /// 3. Apply gain with a `volume={gain_db}dB` graph.
-    /// 4. Store corrected frames in `self.peak_output`.
-    fn run_peak_normalization(&mut self) -> Result<(), FilterError> {
-        let target_db = self
-            .steps
-            .iter()
-            .find_map(|s| {
-                if let FilterStep::NormalizePeak { target_db } = s {
-                    Some(*target_db)
-                } else {
-                    None
-                }
-            })
-            .ok_or(FilterError::BuildFailed)?;
-
-        // Mark done early to prevent re-entry on error.
-        self.peak_pass2_done = true;
-
-        if self.peak_buf.is_empty() {
-            return Ok(());
-        }
-
-        // === Pass 1: measure peak level ===
-        let measured_peak_db = unsafe {
-            let graph = ff_sys::avfilter_graph_alloc();
-            if graph.is_null() {
-                return Err(FilterError::BuildFailed);
-            }
-            let result = run_astats_graph(graph, &self.peak_buf);
-            let mut g = graph;
-            ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(g));
-            result?
-        };
-
-        let gain_db = target_db - measured_peak_db;
-        log::info!(
-            "peak normalization measured_peak_db={:.2} target_db={:.2} gain_db={:.2}",
-            measured_peak_db,
-            target_db,
-            gain_db,
-        );
-
-        // === Pass 2: apply volume correction ===
-        self.peak_output = unsafe {
-            let graph = ff_sys::avfilter_graph_alloc();
-            if graph.is_null() {
-                return Err(FilterError::BuildFailed);
-            }
-            let result = run_volume_graph(graph, &self.peak_buf, gain_db);
-            let mut g = graph;
-            ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(g));
-            result?
-        };
-
-        Ok(())
-    }
-}
-
-// ── EBU R128 two-pass helper functions ───────────────────────────────────────
-
-/// Build a temporary `abuffer → ebur128=peak=true:metadata=1 → abuffersink` graph,
-/// feed all `frames` through it, drain the output, and return the integrated
-/// loudness (LUFS) read from `lavfi.r128.I` on the last output frame.
-///
-/// Falls back to `−70.0` (silence level) if no metadata is found.
-///
-/// # Safety
-///
-/// `graph` must be a valid, freshly-allocated `AVFilterGraph`.  The caller is
-/// responsible for freeing it with `avfilter_graph_free` after this call returns
-/// (whether `Ok` or `Err`).
-unsafe fn run_ebur128_graph(
-    graph: *mut ff_sys::AVFilterGraph,
-    frames: &[AudioFrame],
-) -> Result<f32, FilterError> {
-    let first = &frames[0];
-    let src_args_str = audio_buffersrc_args(
-        first.sample_rate(),
-        sample_format_to_av_name(first.format()),
-        first.channels(),
-    );
-    let src_args = std::ffi::CString::new(src_args_str).map_err(|_| FilterError::BuildFailed)?;
-
-    // 1. abuffersrc
-    let abuffer = ff_sys::avfilter_get_by_name(c"abuffer".as_ptr());
-    if abuffer.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut src_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut src_ctx,
-        abuffer,
-        c"meas_in".as_ptr(),
-        src_args.as_ptr(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // 2. ebur128=peak=true:metadata=1
-    let ebur128_filt = ff_sys::avfilter_get_by_name(c"ebur128".as_ptr());
-    if ebur128_filt.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut meas_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut meas_ctx,
-        ebur128_filt,
-        c"meas_ebur128".as_ptr(),
-        c"peak=true:metadata=1".as_ptr(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // 3. abuffersink
-    let abuffersink = ff_sys::avfilter_get_by_name(c"abuffersink".as_ptr());
-    if abuffersink.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut sink_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut sink_ctx,
-        abuffersink,
-        c"meas_out".as_ptr(),
-        std::ptr::null_mut(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // Link: src → ebur128 → sink
-    let ret = ff_sys::avfilter_link(src_ctx, 0, meas_ctx, 0);
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-    let ret = ff_sys::avfilter_link(meas_ctx, 0, sink_ctx, 0);
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // Configure
-    let ret = ff_sys::avfilter_graph_config(graph, std::ptr::null_mut());
-    if ret < 0 {
-        return Err(ffmpeg_err(ret));
-    }
-
-    // Feed all frames
-    for frame in frames {
-        let raw_frame = ff_sys::av_frame_alloc();
-        if raw_frame.is_null() {
-            return Err(FilterError::ProcessFailed);
-        }
-        (*raw_frame).nb_samples = frame.samples() as c_int;
-        (*raw_frame).sample_rate = frame.sample_rate() as c_int;
-        (*raw_frame).format = sample_format_to_av(frame.format());
-        (*raw_frame).pts = audio_pts_ticks(frame.timestamp(), frame.sample_rate());
-        (*raw_frame).ch_layout.nb_channels = frame.channels() as c_int;
-        let ret = ff_sys::av_frame_get_buffer(raw_frame, 0);
-        if ret < 0 {
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-            return Err(FilterError::ProcessFailed);
-        }
-        copy_audio_planes_to_av(frame, raw_frame);
-        let ret = ff_sys::av_buffersrc_add_frame_flags(
-            src_ctx,
-            raw_frame,
-            ff_sys::BUFFERSRC_FLAG_KEEP_REF,
-        );
-        let mut ptr = raw_frame;
-        ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-        if ret < 0 {
-            return Err(FilterError::ProcessFailed);
-        }
-    }
-
-    // Signal EOF so the filter flushes all pending frames.
-    ff_sys::av_buffersrc_close(src_ctx, ff_sys::AV_NOPTS_VALUE, 0u32);
-
-    // Drain all output; read `lavfi.r128.I` from each frame, keep the last value.
-    let mut last_integrated: f32 = -70.0;
-    loop {
-        let raw_frame = ff_sys::av_frame_alloc();
-        if raw_frame.is_null() {
-            break;
-        }
-        let ret = ff_sys::av_buffersink_get_frame(sink_ctx, raw_frame);
-        if ret < 0 {
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-            break;
-        }
-        // SAFETY: `(*raw_frame).metadata` is a valid `AVDictionary*` (may be null);
-        // `av_dict_get` handles null dictionaries by returning null.
-        let entry = ff_sys::av_dict_get(
-            (*raw_frame).metadata,
-            c"lavfi.r128.I".as_ptr(),
-            std::ptr::null(),
-            0,
-        );
-        if !entry.is_null()
-            && let Ok(s) = std::ffi::CStr::from_ptr((*entry).value).to_str()
-            && let Ok(v) = s.parse::<f32>()
-        {
-            last_integrated = v;
-        }
-        let mut ptr = raw_frame;
-        ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-    }
-
-    Ok(last_integrated)
-}
-
-/// Build a temporary `abuffer → volume={gain_db}dB → abuffersink` graph,
-/// feed all `frames` through it, drain the output, and return the corrected frames.
-///
-/// # Safety
-///
-/// `graph` must be a valid, freshly-allocated `AVFilterGraph`.  The caller is
-/// responsible for freeing it with `avfilter_graph_free` after this call returns
-/// (whether `Ok` or `Err`).
-unsafe fn run_volume_graph(
-    graph: *mut ff_sys::AVFilterGraph,
-    frames: &[AudioFrame],
-    gain_db: f32,
-) -> Result<Vec<AudioFrame>, FilterError> {
-    let first = &frames[0];
-    let src_args_str = audio_buffersrc_args(
-        first.sample_rate(),
-        sample_format_to_av_name(first.format()),
-        first.channels(),
-    );
-    let src_args = std::ffi::CString::new(src_args_str).map_err(|_| FilterError::BuildFailed)?;
-
-    // 1. abuffersrc
-    let abuffer = ff_sys::avfilter_get_by_name(c"abuffer".as_ptr());
-    if abuffer.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut src_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut src_ctx,
-        abuffer,
-        c"vol_in".as_ptr(),
-        src_args.as_ptr(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // 2. volume={gain_db}dB
-    let volume_filt = ff_sys::avfilter_get_by_name(c"volume".as_ptr());
-    if volume_filt.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let vol_args =
-        std::ffi::CString::new(format!("{gain_db:.4}dB")).map_err(|_| FilterError::BuildFailed)?;
-    let mut vol_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut vol_ctx,
-        volume_filt,
-        c"vol_volume".as_ptr(),
-        vol_args.as_ptr(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // 3. abuffersink
-    let abuffersink = ff_sys::avfilter_get_by_name(c"abuffersink".as_ptr());
-    if abuffersink.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut sink_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut sink_ctx,
-        abuffersink,
-        c"vol_out".as_ptr(),
-        std::ptr::null_mut(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // Link: src → volume → sink
-    let ret = ff_sys::avfilter_link(src_ctx, 0, vol_ctx, 0);
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-    let ret = ff_sys::avfilter_link(vol_ctx, 0, sink_ctx, 0);
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // Configure
-    let ret = ff_sys::avfilter_graph_config(graph, std::ptr::null_mut());
-    if ret < 0 {
-        return Err(ffmpeg_err(ret));
-    }
-
-    // Feed all frames
-    for frame in frames {
-        let raw_frame = ff_sys::av_frame_alloc();
-        if raw_frame.is_null() {
-            return Err(FilterError::ProcessFailed);
-        }
-        (*raw_frame).nb_samples = frame.samples() as c_int;
-        (*raw_frame).sample_rate = frame.sample_rate() as c_int;
-        (*raw_frame).format = sample_format_to_av(frame.format());
-        (*raw_frame).pts = audio_pts_ticks(frame.timestamp(), frame.sample_rate());
-        (*raw_frame).ch_layout.nb_channels = frame.channels() as c_int;
-        let ret = ff_sys::av_frame_get_buffer(raw_frame, 0);
-        if ret < 0 {
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-            return Err(FilterError::ProcessFailed);
-        }
-        copy_audio_planes_to_av(frame, raw_frame);
-        let ret = ff_sys::av_buffersrc_add_frame_flags(
-            src_ctx,
-            raw_frame,
-            ff_sys::BUFFERSRC_FLAG_KEEP_REF,
-        );
-        let mut ptr = raw_frame;
-        ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-        if ret < 0 {
-            return Err(FilterError::ProcessFailed);
-        }
-    }
-
-    // Signal EOF
-    ff_sys::av_buffersrc_close(src_ctx, ff_sys::AV_NOPTS_VALUE, 0u32);
-
-    // Drain all corrected output frames
-    let mut output = Vec::new();
-    loop {
-        let raw_frame = ff_sys::av_frame_alloc();
-        if raw_frame.is_null() {
-            break;
-        }
-        let ret = ff_sys::av_buffersink_get_frame(sink_ctx, raw_frame);
-        if ret < 0 {
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-            break;
-        }
-        if let Ok(af) = av_frame_to_audio_frame(raw_frame) {
-            output.push(af);
-        }
-        let mut ptr = raw_frame;
-        ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-    }
-
-    Ok(output)
-}
-
-// ── Peak normalization helper ─────────────────────────────────────────────────
-
-/// Build a temporary `abuffer → astats=metadata=1 → abuffersink` graph,
-/// feed all `frames` through it, drain the output, and return the maximum
-/// peak level (dBFS) read from `lavfi.astats.Overall.Peak_level` across all
-/// output frames.
-///
-/// Falls back to `−70.0` (silence level) if no metadata is found.
-///
-/// # Safety
-///
-/// `graph` must be a valid, freshly-allocated `AVFilterGraph`.  The caller is
-/// responsible for freeing it with `avfilter_graph_free` after this call returns
-/// (whether `Ok` or `Err`).
-unsafe fn run_astats_graph(
-    graph: *mut ff_sys::AVFilterGraph,
-    frames: &[AudioFrame],
-) -> Result<f32, FilterError> {
-    let first = &frames[0];
-    let src_args_str = audio_buffersrc_args(
-        first.sample_rate(),
-        sample_format_to_av_name(first.format()),
-        first.channels(),
-    );
-    let src_args = std::ffi::CString::new(src_args_str).map_err(|_| FilterError::BuildFailed)?;
-
-    // 1. abuffersrc
-    let abuffer = ff_sys::avfilter_get_by_name(c"abuffer".as_ptr());
-    if abuffer.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut src_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut src_ctx,
-        abuffer,
-        c"peak_in".as_ptr(),
-        src_args.as_ptr(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // 2. astats=metadata=1
-    let astats_filt = ff_sys::avfilter_get_by_name(c"astats".as_ptr());
-    if astats_filt.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut meas_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut meas_ctx,
-        astats_filt,
-        c"peak_astats".as_ptr(),
-        c"metadata=1".as_ptr(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // 3. abuffersink
-    let abuffersink = ff_sys::avfilter_get_by_name(c"abuffersink".as_ptr());
-    if abuffersink.is_null() {
-        return Err(FilterError::BuildFailed);
-    }
-    let mut sink_ctx: *mut ff_sys::AVFilterContext = std::ptr::null_mut();
-    let ret = ff_sys::avfilter_graph_create_filter(
-        &raw mut sink_ctx,
-        abuffersink,
-        c"peak_out".as_ptr(),
-        std::ptr::null_mut(),
-        std::ptr::null_mut(),
-        graph,
-    );
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // Link: src → astats → sink
-    let ret = ff_sys::avfilter_link(src_ctx, 0, meas_ctx, 0);
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-    let ret = ff_sys::avfilter_link(meas_ctx, 0, sink_ctx, 0);
-    if ret < 0 {
-        return Err(FilterError::BuildFailed);
-    }
-
-    // Configure
-    let ret = ff_sys::avfilter_graph_config(graph, std::ptr::null_mut());
-    if ret < 0 {
-        return Err(ffmpeg_err(ret));
-    }
-
-    // Feed all frames
-    for frame in frames {
-        let raw_frame = ff_sys::av_frame_alloc();
-        if raw_frame.is_null() {
-            return Err(FilterError::ProcessFailed);
-        }
-        (*raw_frame).nb_samples = frame.samples() as c_int;
-        (*raw_frame).sample_rate = frame.sample_rate() as c_int;
-        (*raw_frame).format = sample_format_to_av(frame.format());
-        (*raw_frame).pts = audio_pts_ticks(frame.timestamp(), frame.sample_rate());
-        (*raw_frame).ch_layout.nb_channels = frame.channels() as c_int;
-        let ret = ff_sys::av_frame_get_buffer(raw_frame, 0);
-        if ret < 0 {
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-            return Err(FilterError::ProcessFailed);
-        }
-        copy_audio_planes_to_av(frame, raw_frame);
-        let ret = ff_sys::av_buffersrc_add_frame_flags(
-            src_ctx,
-            raw_frame,
-            ff_sys::BUFFERSRC_FLAG_KEEP_REF,
-        );
-        let mut ptr = raw_frame;
-        ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-        if ret < 0 {
-            return Err(FilterError::ProcessFailed);
-        }
-    }
-
-    // Signal EOF so the filter flushes all pending frames.
-    ff_sys::av_buffersrc_close(src_ctx, ff_sys::AV_NOPTS_VALUE, 0u32);
-
-    // Drain all output; read `lavfi.astats.Overall.Peak_level` from each frame.
-    // `astats` with no reset accumulates across the whole clip, so the last
-    // frame's metadata holds the overall peak.  We track the maximum across all
-    // frames as a safety net in case the filter resets per frame.
-    let mut max_peak_db: f32 = -70.0;
-    loop {
-        let raw_frame = ff_sys::av_frame_alloc();
-        if raw_frame.is_null() {
-            break;
-        }
-        let ret = ff_sys::av_buffersink_get_frame(sink_ctx, raw_frame);
-        if ret < 0 {
-            let mut ptr = raw_frame;
-            ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-            break;
-        }
-        // SAFETY: `(*raw_frame).metadata` is a valid `AVDictionary*` (may be null);
-        // `av_dict_get` handles null dictionaries by returning null.
-        let entry = ff_sys::av_dict_get(
-            (*raw_frame).metadata,
-            c"lavfi.astats.Overall.Peak_level".as_ptr(),
-            std::ptr::null(),
-            0,
-        );
-        if !entry.is_null()
-            && let Ok(s) = std::ffi::CStr::from_ptr((*entry).value).to_str()
-            && let Ok(v) = s.parse::<f32>()
-        {
-            max_peak_db = max_peak_db.max(v);
-        }
-        let mut ptr = raw_frame;
-        ff_sys::av_frame_free(std::ptr::addr_of_mut!(ptr));
-    }
-
-    Ok(max_peak_db)
-}
-
-impl Drop for FilterGraphInner {
-    fn drop(&mut self) {
-        if let Some(ptr) = self.graph.take() {
-            // SAFETY: `graph` is non-null (guaranteed by `NonNull`), and we are
-            // the sole owner.  `avfilter_graph_free` also frees all
-            // `AVFilterContext`s attached to the graph, so `src_ctxs`,
-            // `vsink_ctx`, and `asink_ctx` must NOT be freed individually.
-            // Filter contexts that held `av_buffer_ref` refs to `hw_device_ctx`
-            // release those refs here as well.
-            unsafe {
-                let mut raw = ptr.as_ptr();
-                ff_sys::avfilter_graph_free(std::ptr::addr_of_mut!(raw));
-            }
-        }
-        // Free our own reference to the hardware device context AFTER the graph
-        // has been freed.  The graph's filter contexts held their own references
-        // (created via `av_buffer_ref` in `create_hw_filter`); those were
-        // released by `avfilter_graph_free` above.
-        if let Some(mut hw_ctx) = self.hw_device_ctx.take() {
-            // SAFETY: `hw_ctx` is the sole remaining reference owned by this
-            // struct; the filter graph (and its filter contexts) has already
-            // been freed above.
-            unsafe {
-                ff_sys::av_buffer_unref(std::ptr::addr_of_mut!(hw_ctx));
-            }
-        }
-    }
 }
 
 // ── Unit tests ────────────────────────────────────────────────────────────────
 
 #[cfg(test)]
 mod tests {
+    use super::build::{audio_buffersrc_args, video_buffersrc_args};
     use super::*;
     use crate::graph::FilterStep;
